+++ run.py
@@ -1,0 +1,3 @@
+from allthegreatshows import app
+
+app.run()
